# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CariTalent API Specification",
        "description": "Dokumentasi API untuk platform Direktori & Booking Talent Ekonomi Kreatif (CariTalent)",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/bookings/{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get Booking by ID",
                "operationId": "4b13c909d897a5edd245741a4aafd9d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/my": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get My Bookings",
                "operationId": "06dd70cd217f92418b8c0405ba44b2da",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/{id}/complete": {
            "put": {
                "tags": [
                    "Booking"
                ],
                "summary": "Complete Booking (EO)",
                "operationId": "2318b676c2c9d6baf840a0ebc09b2a97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/{id}/cancel": {
            "put": {
                "tags": [
                    "Booking"
                ],
                "summary": "Cancel Booking",
                "operationId": "2ad940ac357a1afdada16d2d47bdfabb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/events/{event_id}/recommendations": {
            "get": {
                "tags": [
                    "Matchmaking"
                ],
                "summary": "Get Talent Recommendations",
                "operationId": "db688ae21007bcd18aad83b2272afd0b",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}